refactor: ui/main폴더 추가 및 메인 스크린 분리
