+++ frontend/app/src/main/java/com/markoala/tomoandroid/ui/main/MainScreen.kt
@@ -1,17 +1,10 @@
-package com.markoala.tomoandroid.ui.home
+package com.markoala.tomoandroid.ui.main
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Face
 import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.List
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBar
@@ -24,8 +17,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
@@ -49,7 +40,7 @@ fun BottomNavigationBar(selectedTab: BottomTab, onTabSelected: (BottomTab) -> Un
         NavigationBarItem(
             selected = selectedTab == BottomTab.Friends,
             onClick = { onTabSelected(BottomTab.Friends) },
-            icon = { Icon(Icons.Filled.List, contentDescription = "친구목록") },
+            icon = { Icon(Icons.Filled.Face, contentDescription = "친구목록") },
             label = { Text("친구목록") }
         )
         NavigationBarItem(
@@ -67,73 +58,9 @@ fun BottomNavigationBar(selectedTab: BottomTab, onTabSelected: (BottomTab) -> Un
     }
 }
 
-// 각 탭 화면 예시
-@Composable
-fun HomeTabScreen(paddingValues: PaddingValues) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(paddingValues),
-        contentAlignment = Alignment.Center
-    ) {
-        Text("홈 화면")
-    }
-}
-
-@Composable
-fun FriendsTabScreen(paddingValues: PaddingValues) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(paddingValues),
-        contentAlignment = Alignment.Center
-    ) {
-        Text("친구목록 화면")
-    }
-}
-
-@Composable
-fun ProfileTabScreen(
-    name: String,
-    email: String,
-    userId: String,
-    onSignOut: () -> Unit,
-    paddingValues: PaddingValues
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(paddingValues),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(text = "내 정보 페이지")
-        if (name.isNotEmpty() || email.isNotEmpty()) {
-            Text(text = "이름: $name")
-            Text(text = "이메일: $email")
-            Text(text = "아이디: $userId")
-        }
-        Button(onClick = { onSignOut() }) {
-            Text(text = "로그아웃")
-        }
-    }
-}
-
-@Composable
-fun SettingsTabScreen(paddingValues: PaddingValues) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(paddingValues),
-        contentAlignment = Alignment.Center
-    ) {
-        Text("설정 화면")
-    }
-}
-
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun HomeScreen(onSignOut: () -> Unit) {
+fun MainScreen(onSignOut: () -> Unit) {
     val firebaseAuth = remember { FirebaseAuth.getInstance() }
     val firestore = remember { FirebaseFirestore.getInstance() }
     val user = firebaseAuth.currentUser
@@ -158,10 +85,10 @@ fun HomeScreen(onSignOut: () -> Unit) {
         }
     ) { paddingValues ->
         when (selectedTab) {
-            BottomTab.Home -> HomeTabScreen(paddingValues)
-            BottomTab.Friends -> FriendsTabScreen(paddingValues)
-            BottomTab.Profile -> ProfileTabScreen(name, email, userId, onSignOut, paddingValues)
-            BottomTab.Settings -> SettingsTabScreen(paddingValues)
+            BottomTab.Home -> HomeScreen(paddingValues)
+            BottomTab.Friends -> FriendsScreen(paddingValues)
+            BottomTab.Profile -> ProfileScreen(name, email, userId, onSignOut, paddingValues)
+            BottomTab.Settings -> SettingsScreen(paddingValues)
         }
     }
 }
