+++ frontend/app/src/main/java/com/markoala/tomoandroid/ui/main/ProfileScreen.kt
@@ -1,0 +1,40 @@
+package com.markoala.tomoandroid.ui.main
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ProfileScreen(
+    name: String,
+    email: String,
+    userId: String,
+    onSignOut: () -> Unit,
+    paddingValues: PaddingValues
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(paddingValues),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(text = "내 정보 페이지")
+        if (name.isNotEmpty() || email.isNotEmpty()) {
+            Text(text = "이름: $name")
+            Text(text = "이메일: $email")
+            Text(text = "아이디: $userId")
+        }
+        Button(onClick = { onSignOut() }) {
+            Text(text = "로그아웃")
+        }
+    }
+}
+
